feat(router): Disable routes temporarily
This commit disables the `/events` and `/:pathMatch(.*)*` routes by commenting them out. This is a temporary measure to address issues with authentication and a malfunctioning 404 page. Disabling the routes allows us to focus on resolving the underlying problems without impacting the user experience on other parts of the application.

The `beforeEnter` guard for `/events` was causing unexpected behavior due to issues with local storage authentication. The 404 route was also malfunctioning.

These routes will be re-enabled once the authentication and 404 page issues are resolved.