feat(user-menu): add cursor pointer on user icon hover
This commit enhances the user experience by adding a cursor pointer style to the user icon within the UserMenu component.

The addition of the `hover:cursor-pointer` class provides a visual cue to the user, indicating that the icon is interactive and can be clicked. This small change improves the overall usability of the application by making the user interface more intuitive.