feat: Install Iconify Vue and Radix Icons
This commit introduces Iconify Vue and Radix Icons as dependencies.

The addition of Iconify Vue allows for the easy integration of a wide variety of icons into the application using a consistent and performant approach. Radix Icons, specifically, provides a well-designed and comprehensive set of icons that align with the project's design language.

These new dependencies enhance the visual appeal and user experience by providing a richer set of icons for various UI elements.